requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
argparse
Babel>=1.3
eventlet>=0.17.4
PuLP>=1.0.4
keystonemiddleware>=1.5.0
mox>=0.5.3
Paste
PasteDeploy>=1.5.0
pbr<2.0,>=0.11
posix-ipc
python-keystoneclient>=1.6.0
python-heatclient>=0.3.0
python-muranoclient>=0.5.6
python-novaclient>=2.22.0
python-neutronclient<3,>=2.3.11
python-ceilometerclient>=1.0.13
python-cinderclient>=1.2.2
python-swiftclient>=2.2.0
python-ironicclient>=0.6.0
alembic>=0.7.2
python-glanceclient>=0.18.0
Routes!=2.0,!=2.1,>=1.12.3
six>=1.9.0
oslo.config>=1.11.0 # Apache-2.0
oslo.db>=1.12.0 # Apache-2.0
oslo.policy>=0.5.0 # Apache-2.0
oslo.serialization>=1.4.0 # Apache-2.0
oslo.service>=0.1.0 # Apache-2.0
oslo.utils>=1.6.0 # Apache-2.0
oslo.middleware!=2.0.0,>=1.2.0 # Apache-2.0
oslo.vmware>=0.13.1 # Apache-2.0
oslo.log>=1.2.0                         # Apache-2.0
WebOb>=1.2.3
